% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano-example5.R
\docType{data}
\name{clean5000.df}
\alias{clean5000.df}
\title{Subset of RNAseq gene expression data}
\format{A data.frame with 5 columns and row names.}
\usage{
clean5000.df
}
\description{
A data set containing the results from the analysis of data from RNAseq.
Subset of 5000 genes.
}
\details{
The variables are as follows:
\itemize{
  \item logFC (numeric)
  \item logCPM (numeric)
  \item LR (numeric)
  \item Pvalue (numeric)
  \item outcome (-1, 0, +1)
}
}
\note{
Original data.
}
\keyword{datasets}
