% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr.R
\docType{package}
\name{learnr-package}
\alias{learnr}
\alias{learnr-package}
\title{learnr: "Learn R ...as you learnt your mother tongue" data sets}
\description{
Data sets from the book "Learn R ...as you learnt
your mother tongue" by Pedro J. Aphalo.
}
\details{
Package 'learnr' is a companion to the book "Learn R ...as you learnt your
mother tongue" by Pedro J. Aphalo, available at
\url{http://leanpub.com/learnr}. This package supplies a list of all the
packages used in the book. It also contains all the original data sets used
in the book.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://leanpub.com/learnr}
  \item Report bugs at \url{https://bitbucket.org/aphalo/learnr-pkg/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi}

}
