% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnrbook.R
\docType{package}
\name{learnrbook-package}
\alias{learnrbook}
\alias{learnrbook-package}
\title{learnrbook: Datasets and Code Examples from P. J. Aphalo's "Learn R" Book}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Data, scripts and code from chunks used as examples in the book "Learn R: As a Language" 1ed and 2ed by Pedro J. Aphalo. ISBN 9780367182533 (pbk 1ed); ISBN 9780367182557 (hbk 1ed); ISBN 9780429060342 (ebk 1ed).
}
\details{
Package 'learnr' is a companion to the books Aphalo, P. (2020). Learn R (1ed). New
York: Chapman and Hall/CRC, ISBN 9780367182533" and Aphalo, P. (2023). Learn R (2ed). New
York: Chapman and Hall/CRC, ISBN 9781032516998". This package supplies a list
of all the packages used in the book. It also contains all the original data
sets used in the book as well as code for scripts and code chunks included
in the book. Please, consult the "User Guide" under vignettes for instructions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.r4photobiology.info/learnrbook/}
  \item \url{https://github.com/aphalo/learnrbook-pkg/}
  \item Report bugs at \url{https://github.com/aphalo/learnrbook-pkg/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} (\href{https://orcid.org/0000-0003-3385-972X}{ORCID})

}
