% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_all}
\alias{pkgs_all}
\title{Packages used in book "Learn R: As a Language"}
\format{
A vector of character strings.
}
\usage{
pkgs_all
}
\description{
Character vector of package names for the whole book.
}
\examples{
pkgs_all

}
\keyword{datasets}
