% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_all_1ed}
\alias{pkgs_all_1ed}
\title{Packages used in book "Learn R: As a Language" (1ed)}
\format{
A vector of character strings.
}
\usage{
pkgs_all_1ed
}
\description{
Character vector of package names for the whole book.
}
\examples{
pkgs_all_1ed

}
\keyword{datasets}
