% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_at_github}
\alias{pkgs_at_github}
\title{Packages used in book "Learn R ...as you learnt your mother tongue"}
\format{A vector of character strings.}
\usage{
pkgs_at_github
}
\description{
Character vector of package used in the book, which at the time of writing
are available only from Github.
}
\examples{
pkgs_at_github

}
\keyword{datasets}
