% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch10_2ed}
\alias{pkgs_ch10_2ed}
\title{Packages used in ch. 10 of book "Learn R: As a Language" (2ed)}
\format{
A vector of character strings.
}
\usage{
pkgs_ch10_2ed
}
\description{
Character vector of package names for chapter 10 "Base R and Extensions: Data Sharing ".
}
\examples{
pkgs_ch10_2ed

}
\keyword{datasets}
