% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch7_1ed}
\alias{pkgs_ch7_1ed}
\title{Packages used in ch. 7 of book "Learn R: As a Language" (1ed)}
\format{
A vector of character strings.
}
\usage{
pkgs_ch7_1ed
}
\description{
Character vector of package names for chapter 7 "Grammar of graphics".
}
\examples{
pkgs_ch7_1ed

}
\keyword{datasets}
