% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch8_2ed}
\alias{pkgs_ch8_2ed}
\title{Packages used in ch. 8 of book "Learn R: As a Language" (2ed)}
\format{
A vector of character strings.
}
\usage{
pkgs_ch8_2ed
}
\description{
Character vector of package names for chapter 8 "R Extensions: Data Wrangling".
}
\examples{
pkgs_ch8_2ed

}
\keyword{datasets}
