% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch9_2ed}
\alias{pkgs_ch9_2ed}
\title{Packages used in ch. 9 of book "Learn R: As a Language" (2ed)}
\format{
A vector of character strings.
}
\usage{
pkgs_ch9_2ed
}
\description{
Character vector of package names for chapter 9 "R Extensions: Grammar of Graphics".
}
\examples{
pkgs_ch9_2ed

}
\keyword{datasets}
