% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch_data}
\alias{pkgs_ch_data}
\title{Packages used in book "Learn R ...as you learnt your mother tongue"}
\format{A vector of character strings.}
\usage{
pkgs_ch_data
}
\description{
Character vector of package names for chapter "Storing and manipulating data with R".
}
\examples{
pkgs_ch_data

}
\keyword{datasets}
