% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch_data_io}
\alias{pkgs_ch_data_io}
\title{Packages used in ch. 8 of book "Learn R: As a Language"}
\format{
A vector of character strings.
}
\usage{
pkgs_ch_data_io
}
\description{
Character vector of package names for chapter 8 "Data import and export".
}
\examples{
pkgs_ch_data_io

}
\keyword{datasets}
