% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch_ggplot}
\alias{pkgs_ch_ggplot}
\title{Packages used in book "Learn R ...as you learnt your mother tongue"}
\format{A vector of character strings.}
\usage{
pkgs_ch_ggplot
}
\description{
Character vector of package names for chapter "Plots with ggpplot".
}
\examples{
pkgs_ch_ggplot

}
\keyword{datasets}
