% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch_ggplotext}
\alias{pkgs_ch_ggplotext}
\title{Packages used in book "Learn R ...as you learnt your mother tongue"}
\format{A vector of character strings.}
\usage{
pkgs_ch_ggplotext
}
\description{
Character vector of package names for chapter "Extensions to ggplot".
}
\examples{
pkgs_ch_ggplotext

}
\keyword{datasets}
