% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch_graphics}
\alias{pkgs_ch_graphics}
\title{Packages used in ch. 7 of book "Learn R: As a Language"}
\format{
A vector of character strings.
}
\usage{
pkgs_ch_graphics
}
\description{
Character vector of package names for chapter 7 "Grammar of graphics".
}
\examples{
pkgs_ch_graphics

}
\keyword{datasets}
