% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch_maps_raster}
\alias{pkgs_ch_maps_raster}
\title{Packages used in book "Learn R ...as you learnt your mother tongue"}
\format{A vector of character strings.}
\usage{
pkgs_ch_maps_raster
}
\description{
Character vector of package names for chapter "Plotting maps and images".
}
\examples{
pkgs_ch_maps_raster

}
\keyword{datasets}
