% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used-packages-data.R
\docType{data}
\name{pkgs_ch_performance}
\alias{pkgs_ch_performance}
\title{Packages used in book "Learn R ...as you learnt your mother tongue"}
\format{A vector of character strings.}
\usage{
pkgs_ch_performance
}
\description{
Character vector of package names for chapter "If and when R needs help".
}
\examples{
pkgs_ch_performance

}
\keyword{datasets}
