% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viikki-d29-data.R
\docType{data}
\name{viikki_d29.dat}
\alias{viikki_d29.dat}
\title{Wind direction and speed data}
\format{
A tibble with 3 columns and 1440 rows.
}
\usage{
viikki_d29.dat
}
\description{
A data set containing wind speed and direction measured in Viikki, Helsinki,
Finland with a Vaisala Weather Transmitter WXT530 (sonic anemometer) during
2016-07-29 at 1 min interval.
}
\details{
The variables are as follows:
\itemize{
  \item solar_time (yyyy-mm-dd hh:mm:ss)
  \item WindDir_D1_WVT (degrees)
  \item WindSpd_S_WVT (m/s)
}
}
\examples{
names(viikki_d29.dat)
viikki_d29.dat

}
\references{
P. J. Aphalo, unpublished data.
}
\keyword{datasets}
