% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano-example5.R
\docType{data}
\name{clean5000.df}
\alias{clean5000.df}
\title{Subset of RNAseq gene expression data}
\format{
A data.frame with 5 columns and 5000 row names.
}
\usage{
clean5000.df
}
\description{
A data set containing the results from the analysis of data from RNAseq.
Subset of 5000 genes.
}
\details{
The variables are as follows:
\itemize{
  \item logFC (numeric)
  \item logCPM (numeric)
  \item LR (numeric)
  \item Pvalue (numeric)
  \item outcome (-1, 0, +1)
}
}
\note{
A subset of size 5000 selected by random sampling,
}
\references{
Neha Rai et al. (2020) Plant Cell and Environment,
   https://doi.org/10.1111/pce.13752.
}
\keyword{datasets}
