% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr.R
\docType{package}
\name{learnrbook-package}
\alias{learnrbook}
\alias{learnrbook-package}
\title{learnrbook: Datasets for Aphalo's "Learn R" Book}
\description{
Datasets used in the book "Learn R ...as you learnt
your mother tongue" by Pedro J. Aphalo (2017) <https://leanpub.com/learnr>.
}
\details{
Package 'learnr' is a companion to the book "Learn R ...as you learnt your
mother tongue" by Pedro J. Aphalo, available at
\url{http://leanpub.com/learnr}. This package supplies a list of all the
packages used in the book. It also contains all the original data sets used
in the book.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://leanpub.com/learnr}
  \item \url{http://www.r4photobiology.info/}
  \item Report bugs at \url{https://bitbucket.org/aphalo/learnr-pkg/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi}

}
